services/perception-engine/requirements.txt
# DS3 Perception Engine - Python Dependencies
# Task 2.4: AI Feature Extraction from Photos

# Web Framework
fastapi>=0.109.1
uvicorn[standard]==0.27.0
python-multipart>=0.0.22

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# AI/ML - MediaPipe
mediapipe==0.10.9

# AI/ML - PyTorch & Vision
torch==2.6.0
torchvision==0.21.0

# AI/ML - CLIP
openai-clip==1.0.1
ftfy==6.2.0
regex==2024.5.15
tqdm==4.66.4

# Image Processing
Pillow==10.3.0
opencv-python==4.9.0.80
numpy==1.26.3
scikit-learn==1.4.0
scikit-image==0.22.0

# AWS S3 & KMS
boto3==1.34.0
botocore==1.34.0

# Kafka
kafka-python==2.0.2

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0

# Logging & Monitoring
structlog==24.1.0
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3

# Code Quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0
